requirements.txt
#-r requirements/development.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.4.1
billiard==3.3.0.22
celery==3.1.20
Django==1.9.2
django-admin-bootstrapped==2.5.7
django-authtools==1.4.0
django-braces==1.8.1
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-environ==0.4.0
easy-thumbnails==2.3
kombu==3.0.33
MySQL-python==1.2.5
Pillow==3.1.1
python-dateutil==2.4.2
pytz==2015.7
redis==2.10.5
requests==2.9.1
six==1.10.0
sqlparse==0.1.18
Werkzeug==0.11.4
wheel==0.24.0
